Finished CalculateDamage
-Updated Items script to better work with CalculateDamage
